requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
streamlit>=1.40.0
pydantic>=2.8.0
python-dotenv>=1.0.1
youtube-transcript-api>=0.6.2
pypdf>=5.0.0
beautifulsoup4>=4.12.3
requests>=2.32.3
tiktoken>=0.8.0
